fix US CA Kings County
