+++ src/shared/scrapers/US/CA/kings-county/tests/expected.2020-04-17.json
@@ -1,0 +1,5 @@
+{
+  "cases": 18,
+  "deaths": 1,
+  "recovered": 4
+}
